requirements.txt
langchain
langchain-openai
flask
requests
python-dotenv